apps/git-import/requirements.txt
schedule
pymongo
requests
progressbar2